CheatCodes and MiniGames are Linked
Both CheatCodes and minigames are now connected.

I Mixed both of our systems, Basically I added a int variable to the CheatCode scriptable obj that represents the index of the Minigame list.

SO TLDR Just match the index variable with the saem one on the minigame list and boom they will play that minigame when you enter that cheat.

- need to improve the entering of the cheatcodes as in need to add some extra ways to type a cheat code like typos, different cases etc to make it more consistent.

Still base and has place holder but it works.