Other Cheatcodes
So after lots of work it will take me forever to get Goblin mode, Horse mode, and Sheep to work. I need to move so I made some other cheat codes. The one I added now was max ammo it just replenishes all charges to the players cheat codes. It uses the fistbump mini game. I am next going to add the Heal Cheatcode it will just heal the player not what we hoped for but it will work